Use Vite for dev builds - now with HMR! --- This PR introduces Vite as our solution for running dev builds. It provides an "out-of-the-box" hot module replacing, which should help us build those new UI features much quicker.
There is quite a few changes in this PR, but the gist is that they're quite relevant for making us faster.

There is also a number of "fixes", that align the codebase with the latest syntax, and some improvements like the ability to alias imports, which might be easier to reconcile than relative paths.