Preparation for native notifications support --- This PR includes integration with Google's Firebase Cloud Messaging service (FCM) that will facilitate pushing notifications to our native app.
Currently, native Kotlin code receives the notification if the app is in
the foreground, and passes it onto the UI using the Tauri's event
system.

If the app is in the background or not running, notifications are
handled on the OS service by the notifications manager. How the
notifications will display depends on the channel and group setup within
the Kotlin code that creates a channel for notifications.

This PR also adds iOS preparation for the same plugin that we are
building for Android devices.