Restructured native notification data on the backend --- If notifications do not have "notification" data set, only attached data in a HashMap structure, the Firebase service will not auto handle the recieved notification if the app is in the background or not running.
This PR introduces changes to the FCM data on the backend to denote
specific parts of the info that we require. We've also modified the
native app code to allow for handling of notifications.

As an added bonus, we also have a preparation for a user avatar to be
shown in the notification, for now it's only a placeholder.