Native notifications are displayed and react to current app context --- These changes allow the native android notifications to either be shown or omitted depending on the current app context; and to show the correct context when tapped on (if not already).
For example, this means that if the user is currently within a DM
context, any notification received for their current chat will not be
shown, but as soon as they change their app's context, notifications for
that chat will start appearing. Tapping on the notifications will put
the user back into that DM context.

Since our code is split between UI Svelte and native Kotlin, these have
to communicate for this to work:
- Svelte code reports to native if it's ready to receive events
- If svelte code has not reported ready, events are stored in a queue
- Once UI is ready any queued events are flushed

We are also tracking if the app is in the foreground or background
- If the app is in the background or closed, we immediately display
  any notification via native code
- If the app is in the foreground, notification is sent to UI, UI checks
  if the notification should be shown or not; if yes, UI sends it back
to native to show the notification. This way we have one place in native
code where notifications are constructed.