Bottom menu bar updates --- Updated bottom menu paddings and button sizes. Also updated icons and avatar size.
Added function to report type of the Android navigation, which will
directly affect how much space the app will have at the bottom. If the
device uses button navigation we need more padding, and for gesture
navigation we need less.

Fixed issue with startViewTransition that I saw when inspecting the app
in the android emulator, used the `options` approach for the transition
function arguments as defined in the MDN.

Fixed the app window resizing when the keyboard opens to use the insets
and set window bottom padding instead of it actually resizing. The only
question remains what if the app is split on screen - but that would
probably make it not really useable.

Updated paddings around the message input, though the text input still
needs to be polished.